src/requirements.txt
Django~=3.1
django-cors-headers~=3.2
django-cors-middleware~=1.5
django-createsuperuser
django-extensions~=3.0
django-filter~=2.3
django-redis~=4.12
django-sendfile~=0.3
djangorestframework~=3.11
drf-yasg~=1.17
markdown~=3.0
lxml~=4.0
django_redis~=4.0
pysendfile~=2.0
gunicorn
# "psycopg2
# "mysql
pytz
pycountry
countryinfo~=0.1
timezonefinder~=4.4
iso4217
forex-python~=1.0
Babel~=2.8
Pint~=0.15
networkx~=2.5
sympy~=1.7
requests
channels~=3.0
uncertainties~=3.1
chemicals